Beginnings of `Sys_Exit`.
Based on discussions with BP, we could place a stub that calls `Sys_Exit`
at the bottom of the stack; on program exit from main, this should be popped
into EIP. This stub needs to call an interrupt; if we're calling anything within
`sched` then we're able to be interrupt, the global structures become pre-emptible,
and we run into concurrency issues.

I'm likely going to have to refactor the syscall stuff because BP pushed in an
actual dereferenced struct on the stack as opposed to a pointer to one.
I need the `stack_loc` for the context switch on `Sys_Exit`.