Create fat.c
fat.c is part of the fat 12 file system. This is the c code that defines the struct of a file based on documentation and various functions for setting up reading files and memory.

From what I can tell is that there still needs to be information set up in the kernel and boot loader. I have not done that part yet due to wanting to consult another group member on how we should go about implementing that part.